Rename Startup class to YACReaderHttpServer
